+++ test/dial/test_panel.rb
@@ -71,10 +71,10 @@ module Dial
         assert_select "summary", text: <<-SQL.squish
           SELECT "indicators".* FROM "indicators" WHERE "indicators"."gauge_id" = ? /* Long annotation so the ...
         SQL
-        assert_select "span", text: <<-SQL.squish, count: 4
+        assert_select "span", text: <<-SQL.squish
           SELECT "indicators".* FROM "indicators" WHERE "indicators"."gauge_id" = ? /* Long annotation so the query exceeds the maximum length for presentation and contains newlines */ LIMIT ?
         SQL
-        assert_select "span", text: "+ 5 more queries"
+        assert_select "span", text: "+ 9 more queries"
       end
     end